+++ holo3pos.cpp
@@ -162,7 +162,7 @@ void CreateDisplayImageExample(HANDLE_IPS_ACQ handle_ips, int position_index, in
 
 	//std::chrono::high_resolution_clock::time_point t4 = std::chrono::high_resolution_clock::now();
 	// Start capturing a block of fpg frames
-	tsi::ips::VMemory<uint8_t> buffer(frame_data_size*fpg);
+	tsi::ips::VMemory<uint64_t> buffer(frame_data_size*fpg);
 	//std::chrono::high_resolution_clock::time_point t1 = std::chrono::high_resolution_clock::now();
 
 	CHECK_IPS(IPS_StartGrabbing( handle_ips,
